python/pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ume-logging"
version = "0.0.15"
description = "Uniform JSON logging for University Medicine Essen (UME)"
readme = {file = "README.md", content-type = "text/markdown"}
license-files = ["LICENSE"]
authors = [{name = "Felix Nensa", email = "felix.nensa@uk-essen.de"}]
requires-python = ">=3.9"
dependencies = [
  "python-dateutil>=2.8.2",
  "pytz>=2024.1"
]
[project.optional-dependencies]
fastapi = ["fastapi>=0.110"]
otel = [
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp-proto-http>=1.25.0"
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "fastapi>=0.110",
  "opentelemetry-api>=1.25.0",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp-proto-http>=1.25.0"
]

[project.urls]
Homepage = "https://pypi.org/project/ume-logging/"

[tool.setuptools.packages.find]
where = ["."]               # scan the project root
include = ["umelogging*"]